shoddily integrate the stock code with moop
